+++ hosted/aks/support_matrix/support_matrix_provisioning_test.go
@@ -60,10 +60,14 @@ var _ = Describe("SupportMatrixProvisioning", func() {
 				Expect(err).To(BeNil())
 			})
 			AfterEach(func() {
-				err := helper.DeleteAKSHostCluster(cluster, ctx.RancherClient)
-				Expect(err).To(BeNil())
-				err = helper.DeleteAKSClusteronAzure(clusterName)
-				Expect(err).To(BeNil())
+				if ctx.ClusterCleanup {
+					err := helper.DeleteAKSHostCluster(cluster, ctx.RancherClient)
+					Expect(err).To(BeNil())
+					err = helper.DeleteAKSClusteronAzure(clusterName)
+					Expect(err).To(BeNil())
+				} else {
+					fmt.Println("Skipping downstream cluster deletion: ", clusterName)
+				}
 			})
 
 			It("should successfully provision the cluster", func() {
