+++ hosted/gke/support_matrix/support_matrix_importing_test.go
@@ -61,10 +61,14 @@ var _ = Describe("SupportMatrixImporting", func() {
 				Expect(err).To(BeNil())
 			})
 			AfterEach(func() {
-				err := helper.DeleteGKEHostCluster(cluster, ctx.RancherClient)
-				Expect(err).To(BeNil())
-				err = helper.DeleteGKEClusterOnGCloud(zone, project, clusterName)
-				Expect(err).To(BeNil())
+				if ctx.ClusterCleanup {
+					err := helper.DeleteGKEHostCluster(cluster, ctx.RancherClient)
+					Expect(err).To(BeNil())
+					err = helper.DeleteGKEClusterOnGCloud(zone, project, clusterName)
+					Expect(err).To(BeNil())
+				} else {
+					fmt.Println("Skipping downstream cluster deletion: ", clusterName)
+				}
 			})
 			It("should successfully import the cluster", func() {
 				By("checking cluster name is same", func() {
