+++ hosted/gke/support_matrix/support_matrix_provisioning_test.go
@@ -59,8 +59,12 @@ var _ = Describe("SupportMatrixProvisioning", func() {
 				Expect(err).To(BeNil())
 			})
 			AfterEach(func() {
-				err := helper.DeleteGKEHostCluster(cluster, ctx.RancherClient)
-				Expect(err).To(BeNil())
+				if ctx.ClusterCleanup {
+					err := helper.DeleteGKEHostCluster(cluster, ctx.RancherClient)
+					Expect(err).To(BeNil())
+				} else {
+					fmt.Println("Skipping downstream cluster deletion: ", clusterName)
+				}
 			})
 			It("should successfully provision the cluster", func() {
 				By("checking cluster name is same", func() {
